Butcher.Aromatic

Description

Documentation

newtype Cycle a

A cycle is a list of elements up to cyclic permutation.

Constructors

Cycle [a]

divergence :: (IsDecorated t, Eq (Decoration t), Graded (Decoration t), IsTree t, IsVector t, Num (VectorScalar t), Eq (VectorScalar t), Eq t, Graded t) => PlanarAromatic t -> Vector (VectorScalar t) (PlanarAromatic t)

Compute the divergence of a planar aromatic tree by connecting the root to the vertices.

Examples:

>>> divergence $ paiffb "1[2,3]"
(1 *^ ([(1[2,3])],[]) + 1 *^ ([(1[3],2)],[]) + 1 *^ ([(1[2],3)],[]))_3
>>> divergence $ paiffb "1[2[3],4]"
(1 *^ ([(1[2[3],4])],[]) + 1 *^ ([(1[4],2[3])],[]) + 1 *^ ([(1[4],2,3)],[]) + 1 *^ ([(1[2[3]],4)],[]))_4
>>> divergence $ paiffb "(1),1"
(1 *^ ([(1[1])],[]) + 1 *^ ([(1),(1)],[]))_2
>>> divergence $ paiffb "(1),1[2,3]"
(1 *^ ([(1[1[2,3]])],[]) + 1 *^ ([(1[2,3]),(1)],[]) + 1 *^ ([(1[3],2),(1)],[]) + 1 *^ ([(1[2],3),(1)],[]))_4
>>> divergence $ paiffb "1[2,3],4,(5)"
(1 *^ ([(5)],[4[1[2,3]]]) + 1 *^ ([(5[1[2,3]])],[4]) + 1 *^ ([(1[2,3]),(5)],[4]) + 1 *^ ([(1[3],2),(5)],[4]) + 1 *^ ([(1[2],3),(5)],[4]))_5

type PlanarAromatic t = ([Cycle t], [t])

type Aromatic t = (MultiSet (Cycle t), MultiSet t)

data Marked a

We note that a cycle is obtained as a trace of marked tree which replaces the Mark by an incoming edge from the root. A marked tree is obtained from a decorated tree by adding a special Mark decoration to the set of decorations.

Let a be the set of decorations of a tree, then Marked a is the same set with an additional element Mark.

See Fløystad, G., Manchon, D., & Munthe-Kaas, H. Z. (2021). The universal pre-Lie–Rinehart algebras of aromatic trees. https://doi.org/10.1007/978-3-030-78346-4_9

Constructors

Marked a

mark :: Markable a => a -> Marked' a

unmark :: Markable a => Marked' a -> a