Butcher.Graph

Description

Implementation of the general Graph and RootedGraph typeclasses. As an example, the IntegerGraph type is defined as an instance of Graph. The data Rooted is defined as a map from Graph to RootedGraph. The graftGraph and graftGraphTo functions are also defined here.

Documentation

class Graph g

A graph g must have an instance of the Graph typeclass with the edge type Edge g being an instance of the GraphEdge typeclass and the vertex type Vertex g being the EndPoint of Edge g.

Minimal complete definition

singleton, edges, vertices, addGraph, addEdge

type family Vertex g

The type of the vertices of the graph.

singleton :: Graph g => Vertex g -> g

A graph with a single vertex and no edges.

edges :: Graph g => g -> MultiSet (Vertex g, Vertex g)

The set of edges of the graph.

vertices :: Graph g => g -> MultiSet (Vertex g)

The set of vertices of the graph.

addGraph :: (Graph g, Graph g0, Vertex g ~ Vertex g0) => g0 -> g -> g

Add a graph to another graph.

addEdge :: Graph g => (Vertex g, Vertex g) -> g -> g

Add an edge to a graph.

data IntegerGraph

Naive implementation of a graph given by a multiset of vertices and a multiset of edges.

integerGraph :: [Integer] -> [(Integer, Integer)] -> IntegerGraph

A constructor of the IntegerGraph type.

Example:

>>> integerGraph [1, 2, 3] [(1, 2), (2, 3)]
IntegerGraph(V=[1,2,3], E=[(1,2),(2,3)])

class Graph g => RootedGraph g

A typeclass for rooted graphs which have a distinguished vertex called the root.

Minimal complete definition

root

root :: RootedGraph g => g -> Vertex g

data Rooted g

A data type for rooted graphs.

rooted :: (Graph g, Eq (Vertex g)) => g -> Vertex g -> Rooted g

A constructor for the Rooted type which checks if the root vertex is in the graph.

Example:

>>> rooted (integerGraph [1, 2, 3] [(1, 2), (2, 3)]) 1
RootedIntegerGraph(V=[1,2,3], E=[(1,2),(2,3)], R=1)

graftGraph :: (Eq a2, Graded a2, RootedGraph a1, Graph a2, Vertex a1 ~ Vertex a2) => a1 -> a2 -> Vector Integer a2

Grafing of two graphs.

Example:

>>> g1 = integerGraph [1, 2, 3] [(2, 1), (3, 2)]
>>> graftGraph (rooted g1 1) (integerGraph [4, 5] [(5, 4)])
(1 *^ IntegerGraph(V=[1,2,3,4,5], E=[(1,4),(2,1),(3,2),(5,4)]) + 1 *^ IntegerGraph(V=[1,2,3,4,5], E=[(1,5),(2,1),(3,2),(5,4)]))_5

graftGraphTo :: (RootedGraph a1, Graph a2, Vertex a1 ~ Vertex a2) => a1 -> a2 -> Vertex a2 -> a2

Grafing of a rooted graph to a graph at a given vertex.

Example:

>>> g1 = integerGraph [1, 2, 3] [(2, 1), (3, 2)]
>>> graftGraphTo (rooted g1 1) (integerGraph [4, 5] [(5, 4)]) 5
IntegerGraph(V=[1,2,3,4,5], E=[(1,5),(2,1),(3,2),(5,4)])

getVertex :: State [a] a

Get the first vertex of a list of vertices and remove it from the list. This function is used to build graphs from a list of vertices using the State monad.

In practice, it is used to build graphs with unique vertices that are collected from a list of vertices.

Example:

>>> runState getVertex [1, 2, 3]
(1,[2,3])
>>> evalState (do {v1 <- getVertex; v2 <- getVertex; return (integerGraph [v1, v2] [(v2, v1)])}) [1, 2, 3]
IntegerGraph(V=[1,2], E=[(2,1)])