Butcher.NonPlanar

Description

Implementation of the post-Lie algebra of planar trees and pre-Lie algebra of non-planar trees.

Documentation

data Tree d

Non-planar trees are represented as a tree with a root and a multiset of branches which are non-planar trees themselves.

Constructors

T

Fields

nonplanarRoot :: d

nonplanarBranches :: MultiSet (Tree d)

class Planarable t where

Associated Types

type Planar t

Methods

nonplanar :: Planar t -> t

planar :: t -> Planar t

Orphan instances

(Ord t, IsVector t, IsTree t) => CanConnesKreimer (MultiSet t)

Instance details

Methods

connesKreimer :: MultiSet t -> Vector (VectorScalar (MultiSet t)) (VectorBasis (MultiSet t), VectorBasis (MultiSet t))