Butcher.Planar

Description

See Munthe-Kaas, H. Z., & Føllesdal, K. K. (2018). Lie–Butcher Series, Geometry, Algebra and Computation. https://doi.org/10.1007/978-3-030-01397-4_3

Magmatic product and the related formulas are introduced in Section 4 of the paper referenced above.

We implement only those formulas which are clearly "better" than the alternatives which do not use the magmatic product. In particular, the magmatic product gives nice formulas for: - the planar factorial, - planar Connes-Kreimer coproduct, - planar Connes-Kreimer antipode, - planar pruning, - cosubstitution.

Documentation

data PlanarTree d

Planar trees are represented as a tree with a root and a list of branches which are planar trees themselves.

Constructors

PT

Fields

planarRoot :: d

planarBranches :: [PlanarTree d]

magProduct :: IsTree t => Decoration t -> [t] -> [t] -> [t]

Given two planar forests f1 and f2, the magmatic product of f1 and f2 is the planar forest obtained by concatenation of f1 and the tree obtained by grafting all trees in f2 to a new root decorated by c.

Examples:

>>> f1 = iffb "1[2],3"
>>> f2 = iffb "4,5[6]"
>>> magProduct 7 f1 f2
[1[2],3,7[4,5[6]]]

magLeft :: [t] -> [t]

Every planar forest can be uniquely decomposed into a left part, a right part, and a root decoration.

Examples:

>>> magLeft $ iffb "1[2,3],4[5,6]"
[1[2,3]]
>>> magRight $ iffb "1[2,3],4[5,6]"
[5,6]
>>> magRoot $ iffb "1[2,3],4[5,6]"
4

magRight :: IsTree t => [t] -> [t]

magRoot :: IsTree t => [t] -> Decoration t

planarFactorial :: IsTree t => [t] -> Integer

Planar factorial which is the coefficient map for the Lie-Butcher series of the exact solution.

See Equation 47 in Munthe-Kaas, H. Z., & Føllesdal, K. K. (2018). Lie–Butcher Series, Geometry, Algebra and Computation. https://doi.org/10.1007/978-3-030-01397-4_3

Examples:

>>> f = iffb "1[2,3[4]]"
>>> planarFactorial f
12

shuffleCross :: (IsTree t, IsVector t) => Decoration t -> ([t], [t]) -> ([t], [t]) -> Vector (VectorScalar t) ([t], [t])

A product on the tensor algebra of planar forests. Applies shuffle product to the left parts and magmatic product to the right parts. Used internally in the implementation of the Connes-Kreimer coproduct and the pruning formula.

prune :: (IsTree t, IsVector t) => [t] -> Vector (VectorScalar t) ([t], [t])

Planar pruning coproduct dual to the grafting product.

See Equations 68 and 69 in Munthe-Kaas, H. Z., & Føllesdal, K. K. (2018). Lie–Butcher Series, Geometry, Algebra and Computation. https://doi.org/10.1007/978-3-030-01397-4_3

Examples:

>>> f = iffb "1[2,3[4]]"
>>> prune f
(1 *^ ([2,3[4]],[1]) + 1 *^ ([2,4],[1[3]]) + 1 *^ ([4,2],[1[3]]) + 1 *^ ([4],[1[2,3]]) + 1 *^ ([2],[1[3[4]]]) + 1 *^ ([],[1[2,3[4]]]))_4

planarAntipodeCK :: (IsTree t, IsVector t) => [t] -> Vector (VectorScalar t) [t]

Planar Connes-Kreimer antipode.

See Equation 71 in Munthe-Kaas, H. Z., & Føllesdal, K. K. (2018). Lie–Butcher Series, Geometry, Algebra and Computation. https://doi.org/10.1007/978-3-030-01397-4_3

Examples:

>>> f = iffb "1[1,1[1]]"
>>> planarAntipodeCK f
(12 *^ [1,1,1,1] + -2 *^ [1,1,1[1]] + -3 *^ [1,1[1],1] + -4 *^ [1[1],1,1] + 1 *^ [1,1[1,1]] + 1 *^ [1[1,1],1] + 1 *^ [1,1[1[1]]] + 1 *^ [1[1[1]],1] + -1 *^ [1[1,1[1]]])_4

cosub :: (IsTree t, IsVector t) => [Decoration t] -> (Decoration t -> [t] -> VectorScalar t) -> [t] -> Vector (VectorScalar t) [t]

Cosubstitution.

(!!!) The functional a is assumed to satisfy a c [] = 0 and a c (shuffle f1 f2) = 0.

See Equation 79 in Munthe-Kaas, H. Z., & Føllesdal, K. K. (2018). Lie–Butcher Series, Geometry, Algebra and Computation. https://doi.org/10.1007/978-3-030-01397-4_3

Examples:

>>> f = iffb "1[1,1[1]]"
>>> a _ = \case [t] -> grading t; _ -> 0
>>> cosub [1] a f
(4 *^ [1])_1 + (6 *^ [1[1]])_2 + (6 *^ [1[1,1]])_3 + (1 *^ [1[1,1[1]]])_4

buildSyntacticTree :: (IsTree t, IsVector t, Show (Decoration t)) => [t] -> SyntacticTree [t]

Orphan instances

(IsTree t, IsVector t) => CanConnesKreimer [t]
Planar Connes-Kreimer coproduct dual to the Grossman-Larson product.

See Equations 66 and 67 in Munthe-Kaas, H. Z., & Føllesdal, K. K. (2018). Lie–Butcher Series, Geometry, Algebra and Computation. https://doi.org/10.1007/978-3-030-01397-4_3

Examples:

>>> f = iffb "1[2,3[4]]"
>>> connesKreimer f
(1 *^ ([1[2,3[4]]],[]) + 1 *^ ([2,3[4]],[1]) + 1 *^ ([2,4],[1[3]]) + 1 *^ ([4,2],[1[3]]) + 1 *^ ([4],[1[2,3]]) + 1 *^ ([2],[1[3[4]]]) + 1 *^ ([],[1[2,3[4]]]))_4

Instance details

Methods

connesKreimer :: [t] -> Vector (VectorScalar [t]) (VectorBasis [t], VectorBasis [t])