Butcher.Series

Description

(EXPERIMENTAL) Implementation of Butcher and Lie-Butcher series over classical, decorated, and aromatic forests

Documentation

expGL :: (IsTree t, IsVector t, Eq (Decoration t), Graded (Decoration t), Fractional (VectorScalar t)) => Vector (VectorScalar t) [t] -> Vector (VectorScalar t) [t] -> Vector (VectorScalar t) [t]

Consider an ODE $dy/dt = f(y)$ and let $f(y)$ be expressed using a linear combination of trees $v$. Let $phi circ y$ be the exact solution composed with a test function $phi$, then, \[ \phi \circ y = \exp^{GL} (v) [\phi] , \] where $exp^{GL}$ is the Grossman-Larson exponential and the square brackets denote the application of the corresponding differential operator to the test function $phi$.

The function expGL implements the exponential $exp^{GL}$.

bSeries :: (IsTree t, IsVector t, Eq (Decoration t), Graded (Decoration t)) => [MultiSet t] -> (MultiSet t -> VectorScalar t) -> Vector (VectorScalar t) (MultiSet t)

Construct a Butcher series over non-ordered forests. The function bSeries takes a non-decreasing list of forests and a function that assigns a coefficient to each forest, and returns the corresponding Butcher series as a vector.

lbSeries :: (IsTree t, IsVector t, Eq (Decoration t), Graded (Decoration t)) => [[t]] -> ([t] -> VectorScalar t) -> Vector (VectorScalar t) [t]

Construct a Lie-Butcher series over ordered forests. The function lbSeries takes a non-decreasing list of forests and a function that assigns a coefficient to each forest, and returns the corresponding Lie-Butcher series as a vector.

rkCoeff :: IsTree t => Int -> Matrix R -> Vector R -> t -> R

Compute the Runge-Kutta coefficient for a given tree. The function rkCoeff takes the number of stages s, the matrix of coefficients aij, the vector of weights bi, and a tree t, and returns the corresponding Runge-Kutta coefficient as a real number.

rkBSeries :: (IsTree t, IsVector t, Eq (Decoration t), Graded (Decoration t), VectorScalar t ~ R, Fractional (VectorScalar t)) => [MultiSet t] -> Int -> Matrix R -> Vector R -> Vector (VectorScalar t) (MultiSet t)

Construct a Butcher series over non-ordered forests listed in forests of a Runge-Kutta method with s stages and coefficients bi and aij.