Butcher.Tree

Description

Define the class of decorated trees and some common functions for all tree structures. The primary example of a decorated tree is the planar decorated tree implemented in the Butcher.Planar module, but the functions in this module are designed to be as general as possible and can be used for other tree structures as well.

Documentation

class IsDecorated a

A type a is decorated if it has IsDecorated instance with type Decoration a being the type of the decoration.

Associated Types

type Decoration a

class (IsDecorated t, Graded t) => IsTree t where

A decorated tree is a decorated and graded structure with a root and a list of branches. The grading is necessary to define the corresponding graded vector space.

Methods

root :: t -> Decoration t

branches :: t -> [t]

buildTree :: Decoration t -> [t] -> t

class IsDecorated t => HasBracketNotation t where

A decorated structure t can be written and constructed from a string by implementing an instance of the HasBracketNotation class.

Minimal complete definition

toBracketsWith, fromBracketsWith

Methods

toBracketsWith :: (Decoration t -> String) -> t -> String

Convert a t to a string using bracket notation. The first argument is a function that converts the decoration to a string.

toBrackets :: Show (Decoration t) => t -> String

fromBracketsWith :: (String -> Decoration t) -> String -> t

Convert a string to t using bracket notation.

fromBrackets :: Read (Decoration t) => String -> t

treeToBracketsWith :: IsTree t => (Decoration t -> String) -> t -> String

Every tree can be written and constructed from a string using the bracket notation. This is the default impelementation of the toBracketsWith and fromBracketsWith functions for any IsTree t.

See IsTree (PlanarTree d) instance in Planar module for examples.

treeFromBracketsWith :: IsTree t => (String -> Decoration t) -> String -> t

treeTexify :: (Show (Decoration t), IsTree t) => t -> String

We use planarforest PythonTeX package for generating the TeX code for trees. The function treeTexify converts a tree to a string format accepted by the planarforest package.

parseTree :: IsTree t => (String -> Decoration t) -> State String t

The function parseTree is used to parse a string into a tree using the bracket notation.

See IsTree (PlanarTree d) instance in Planar module for examples.

parseForest :: IsTree t => (String -> Decoration t) -> State String [t]

The function parseForest is used to parse a string into a forest using the bracket notation.

See IsTree (PlanarTree d) instance in Planar module for examples.

parseDecoration :: (String -> d) -> State String d

The function parseDecoration is used to parse a string as a decoration using the bracket notation.

Example:

>>> evalState (parseDecoration read) "1234[" :: Integer
1234

class IsVector a => CanGraft a where

Methods

graft :: a -> a -> Vector (VectorScalar a) (VectorBasis a)

class IsVector a => CanGrossmanLarson a where

Methods

grossmanLarson :: a -> a -> Vector (VectorScalar a) (VectorBasis a)

class IsVector a => CanConnesKreimer a where

Methods

connesKreimer :: a -> Vector (VectorScalar a) (VectorBasis a, VectorBasis a)

graftOp :: (IsVector a, CanGraft a) => Operation a

Grafting operation.

concatOp :: (IsVector a, Monoid a) => Operation a

Concatenation operation.

Orphan instances

(IsVector t, IsTree t, Ord t) => HasSyntacticTree (MultiSet t)
Construct a syntactic tree of a multiset of trees.

Instance details

Methods

syn :: MultiSet t -> SyntacticTree (MultiSet t)

(IsVector t, IsTree t) => HasSyntacticTree [t]
Construct a syntactic tree of a list of trees.

See IsTree (PlanarTree d) instance in Planar module for examples.

Instance details

Methods

syn :: [t] -> SyntacticTree [t]