Core.Algebra

Description

Documentation

class IsVector a => CanDeshuffle a where

a is assumed to be the basis element.

Methods

deshuffle :: a -> Vector (VectorScalar a) (a, a)

shuffle :: (IsVector v, Graded v) => [v] -> [v] -> Vector (VectorScalar v) [v]

Shuffle product of lists.

Examples:

>>> shuffle "xy" "uv"
(1 *^ "xyuv" + 1 *^ "xuyv" + 1 *^ "xuvy" + 1 *^ "uxyv" + 1 *^ "uxvy" + 1 *^ "uvxy")_4

deconcatenate :: (IsVector v, Graded v) => [v] -> Vector (VectorScalar v) ([v], [v])

Deconcatenation coproduct for lists.

Examples:

>>> deconcatenate "xyz"
(1 *^ ("","xyz") + 1 *^ ("x","yz") + 1 *^ ("xy","z") + 1 *^ ("xyz",""))_3

antipode :: (IsVector v, Graded v) => [v] -> Vector (VectorScalar v) [v]

Antipode of the Concatenate-Deshuffle and Shuffle-Deconcatenate Hopf algebras.

Examples:

>>> a = "xyz"
>>> antipode a
(-1 *^ "zyx")_3

antipodeWith :: (Graded a, IsVector a, VectorBasis a ~ a) => a -> (a -> a -> Vector (VectorScalar a) a) -> (a -> Vector (VectorScalar a) (a, a)) -> a -> Vector (VectorScalar a) a

Antipode of a combinatorial HOpf algebra defined by a unit, a product, and a coproduct. It implements the recursive formula for an antipode of a connected graded bialgebra. It is highly inefficient and should be used only for exploratory purposes.

Examples:

>>> a = "xyz"
>>> concat x y = vector $ 1 *^ (x ++ y)
>>> antipodeWith [] concat deshuffle "xyz"
(-1 *^ "zyx")_3

morphism :: (Functor f, Foldable f, Eq (f a), Monoid b, IsVector v, VectorScalar v ~ k, VectorBasis v ~ b) => (a -> v) -> Vector k (f a) -> Vector k b

Take a function f that maps basis elements to basis elements and extends it to a morphism of the algebra.

!!! The function morphism f accepts only finite vectors.

Examples:

>>> f1 x = if x < 9 then 1 *^ [x+1] else 1 *^ [x, 1]
>>> v1 = vectorFromList [1 *^ [6, 7], 1 *^ [8, 9]]
>>> morphism f1 v1
(1 *^ [7,8])_2 + (1 *^ [9,9,1])_3
>>> f2 x = vectorFromList [1 *^ [x], 1 *^ [x+1]]
>>> v2 = vectorFromList [1 *^ [1, 2], 1 *^ [3, 4]]
>>> morphism f2 v2
(1 *^ [1,2] + 1 *^ [2,2] + 1 *^ [1,3] + 1 *^ [2,3] + 1 *^ [3,4] + 1 *^ [4,4] + 1 *^ [3,5] + 1 *^ [4,5])_2

morphismGraded :: (Functor f, Foldable f, Eq (f a), Monoid b, IsVector v, VectorScalar v ~ k, VectorBasis v ~ b) => (a -> v) -> Vector k (f a) -> Vector k b

Take a function f that maps basis elements to basis elements and extends it to a morphism of the tensor algebra.

!!! The function f must preserve the grading.

Examples:

>>> f1 x = 1 *^ [1,x]
>>> v1 = vectorFromList [1 *^ [6,7], 1 *^ [8, 9]]
>>> morphismGraded f1 v1
*** Exception: Grading mismatch in a list of terms
...
>>> f2 x = 1 *^ [x + 1]
>>> v2 = vectorFromList [1 *^ [6, 7], 1 *^ [8, 9]]
>>> morphismGraded f2 v2
(1 *^ [7,8] + 1 *^ [9,10])_2
>>> v3 = vectorFromNonDecList [i *^ (replicate i 1) | i <- [1..]]
>>> takeV 10 $ morphismGraded f2 v3
(1 *^ [2])_1 + (2 *^ [2,2])_2 + (3 *^ [2,2,2])_3 + (4 *^ [2,2,2,2])_4 + (5 *^ [2,2,2,2,2])_5 + (6 *^ [2,2,2,2,2,2])_6 + (7 *^ [2,2,2,2,2,2,2])_7 + (8 *^ [2,2,2,2,2,2,2,2])_8 + (9 *^ [2,2,2,2,2,2,2,2,2])_9 + (10 *^ [2,2,2,2,2,2,2,2,2,2])_10

morphismNonDec :: (Functor f, Foldable f, Eq (f a), Monoid b, IsVector v, VectorScalar v ~ k, VectorBasis v ~ b) => (a -> v) -> Vector k (f a) -> Vector k b

Take a function f that maps basis elements to basis elements and extends it to a morphism of the tensor algebra.

!!! The function f must be non-decreasing (see linearNonDec).

Examples:

>>> f1 x = 1 *^ [x+1]
>>> v1 = vectorFromList [1 *^ [6,7], 1 *^ [8, 9]]
>>> morphismNonDec f1 v1
(1 *^ [7,8] + 1 *^ [9,10])_2
>>> f2 x = if x > 0 && x < 9 then 1*^ [x + 1] else 1 *^ [x]
>>> v2 = vectorFromList [1 *^ [6, 7], 1 *^ [8, 9]]
>>> morphismNonDec f2 v2
(1 *^ [7,8] + 1 *^ [9,9])_2
>>> v3 = vectorFromNonDecList [i *^ (replicate i 1) | i <- [1..]]
>>> takeV 10 $ morphismNonDec f2 v3
(1 *^ [2])_1 + (2 *^ [2,2])_2 + (3 *^ [2,2,2])_3 + (4 *^ [2,2,2,2])_4 + (5 *^ [2,2,2,2,2])_5 + (6 *^ [2,2,2,2,2,2])_6 + (7 *^ [2,2,2,2,2,2,2])_7 + (8 *^ [2,2,2,2,2,2,2,2])_8 + (9 *^ [2,2,2,2,2,2,2,2,2])_9 + (10 *^ [2,2,2,2,2,2,2,2,2,2])_10