Core.GradedList

Description

Implementation of a Graded typeclass and utilities for lists of graded elements used in Symbolics module.

Documentation

class Graded t

The Graded typeclass is used to define the grading of an element.

Minimal complete definition

grading

grading :: Graded t => t -> Integer

data NonDecreasingList a

A list in which the gradings of the elements is non-decreasing.

unNDecList :: NonDecreasingList a -> [a]

nDecList :: Graded a => [a] -> NonDecreasingList a

Construct a graded list by checking that the grading is non-decreasing.

Example:

>>> nDecList ["x", "xy", "xy", "xyz"]
NonDecreasingList ["x","xy","xy","xyz"]
>>> nDecList ["x", "xy", "xyz", "xy"]
NonDecreasingList ["x","xy"*** Exception: The list given to @gradedList@ is not graded.
...

distributeLists :: Eq a => [[a]] -> [[a]]

Cartesian product of lists in which lists can be infinite. It works in the following way: distributeLists [[a_1, a_2], [b_1, b_2]] will return: [[a_1, b_1], [a_1, b_2], [a_2, b_1], [a_2, b_2]].

Example:

>>> distributeLists [[1, 2], [11, 12], [21, 22]]
[[1,11,21],[2,11,21],[1,12,21],[1,11,22],[2,12,21],[2,11,22],[1,12,22],[2,12,22]]

distributeGradedLists :: Eq a => [[a]] -> [[[a]]]

The same as distributeLists but it groups the resulting terms by the sums of their indices. It works in the following way: distributeLists [[a_1, a_2, a_3], [b_1, b_2], [c_1, c_2]] will return: [[[a_1,b_1,c_1]],[[a_2,b_1,c_1],[a_1,b_2,c_1],[a_1,b_1,c_2]], [[a_3,b_1,c_1],[a_2,b_2,c_1],[a_2,b_1,c_2],[a_1,b_2,c_2]], [[a_3,b_2,c_1],[a_3,b_1,c_2],[a_2,b_2,c_2]],[[a_3,b_2,c_2]]].

Example:

>>> distributeGradedLists [[1, 2, 3], [11, 12], [21, 22]]
[[[1,11,21]],[[2,11,21],[1,12,21],[1,11,22]],[[3,11,21],[2,12,21],[2,11,22],[1,12,22]],[[3,12,21],[3,11,22],[2,12,22]],[[3,12,22]]]

groupByGrading :: Graded a => NonDecreasingList a -> [HomoList a]

Same as groupByGradingFrom but uses the grading of the elements and starts with grading 0.

Example:

>>> groupByGrading $ nDecList ["x", "y", "xy", "xy", "yz", "zx", "xyz", "zxy"]
[[],["x","y"],["xy","xy","yz","zx"],["xyz","zxy"]]