Core.Output

Description

Implementation of the Texifiable typeclass, its instances for the Symbolics module and functions to display the output in a pdf viewer.

Documentation

class Texifiable a

To be displayed in a pdf using LaTeX, a type must be an instance of the Texifiable typeclass.

It is mandatory to define the texify function. You can specify the parentheses of the type by defining texifyParentheses. You can also define texifyDebug to display debug information and texifyID to display the identifier of the type.

Minimal complete definition

texify

texify :: Texifiable a => a -> String

The main function of the typeclass. It returns the TeX code of a. Use texifyP for recursion when defining this if you want the parentheses to be displayed.

texifyParentheses :: Texifiable a => a -> (String, String)

Returns the TeX code of the left and right parentheses of a. By default, it returns the empty string for both.

texifyDebug :: Texifiable a => Integer -> Integer -> a -> String

Returns the TeX code of a with debug information. Use texifyD for recursion when defining this. By default, it returns the same as texifyP. The first two arguments are used to specify the levels of the syntax tree at which the debug information is displayed.

texifyID :: Texifiable a => a -> String

The identifier of the type used for debugging. By default, it returns the empty string.

display :: (IsVector v, Texifiable (VectorScalar v), Texifiable (VectorBasis v)) => v -> IO ()

Generate a TeX string from a type and display it in a pdf viewer.

displayDebug :: (IsVector v, Texifiable (VectorScalar v), Texifiable (VectorBasis v)) => Integer -> Integer -> v -> IO ()

Generate a TeX string with debug information from a type and display it in a pdf viewer.