Core.SyntacticTree

Description

Implementation of syntactic trees is used to represent elements of an algebra as expressions involving generators and operations. This allows us to use syntactic trees to define automorphisms and, if extended, homomorphisms.

Documentation

data Operation a

An operation is a function that takes a list of operands and returns a vector. The value of arity is the number of operands accepted by the operation. The func performs the actual operation. The tex is the LaTeX representation of the operation.

!!! The name of the operation must be unique.

Constructors

Op

Fields

name :: String

tex :: String

arity :: Int

func :: [a] -> Vector (VectorScalar a) (VectorBasis a)

data SyntacticTree a

Syntactic tree is a tree with operations at the nodes and operands at the leaves.

Constructors

Node (Operation a) [SyntacticTree a]

Leaf a

compose :: (Eq a, Graded a) => SyntacticTree a -> [SyntacticTree a] -> SyntacticTree a -> Maybe (SyntacticTree a)

Replace the operands which are equal to the syntactic tree x of a syntactic tree y with the given list of syntactic trees ops. The operands are replaced in the order they appear.

The function returns Nothing if the number of syntactic trees in the list ops is different from the number of operands in the syntactic tree y.

See IsTree (PlanarTree d) instance in Planar module for examples.

symmetricCompose :: (IsVector a, Graded a) => SyntacticTree a -> [SyntacticTree a] -> SyntacticTree a -> Vector (VectorScalar (SyntacticTree a)) (SyntacticTree a)

The same as compose but we sum over all permutations of the list of syntactic trees and return the result as a vector.

See IsTree (PlanarTree d) instance in Planar module for examples.

class IsVector a => HasSyntacticTree a where

A class for types that can be represented as syntactic trees.

Methods

syn :: a -> SyntacticTree a

eval :: (IsVector a, VectorBasis a ~ a) => SyntacticTree a -> Vector (VectorScalar a) a

Evaluate a syntactic tree.

See IsTree (PlanarTree d) instance in Planar module for examples.

substitute :: (HasSyntacticTree a, IsVector a, VectorBasis a ~ a) => a -> [a] -> a -> Vector (VectorScalar a) a

Substitute the subexpression x of obj with the given list of expressions gens in all possible ways and return the result as a vector.

If the number of expressions in the list gens is different from the number of occurrences of the subexpression x in obj, the function returns a zero vector.

See IsTree (PlanarTree d) instance in Planar module for examples.