Core.VectorSpace

Description

An implementation of symbolic algebra using graded vector spaces with the aim of being able to represent and manipulate algebras over the vector spaces of graphs.

Documentation

data ScalarProduct k a

A scalar product is defined to be a pair of a scalar (Num) and a basis element.

scalar :: ScalarProduct k a -> k

basisElement :: ScalarProduct k a -> a

pattern (:*^) :: k -> a -> ScalarProduct k a

(*^) :: Num k => k -> a -> ScalarProduct k a infix 7

scaleSP :: Num k => k -> ScalarProduct k a -> ScalarProduct k a

Graded vector space

class (Eq v, Eq (VectorBasis v), Graded (VectorBasis v), Eq (VectorScalar v), Num (VectorScalar v)) => IsVector v where

A class of types that can be cast to a vector.

Associated Types

type VectorScalar v

type VectorBasis v

Methods

vector :: v -> Vector (VectorScalar v) (VectorBasis v)

newtype Finite v

Some functions assume the vector to be finite.

Constructors

Finite

Fields

getFinite :: v

vectorFromList :: (Num k, Eq k, Eq a, Graded a) => [ScalarProduct k a] -> Vector k a

Construct a vector from a finite list of terms by ordering the terms according to the gradings.

!!! The list must be finite.

Examples:

>>> vectorFromList [1 *^ "x", 1 *^ "y", 3 *^ "xy", 1 *^ "x", 1 *^ "y"]
(2 *^ "x" + 2 *^ "y")_1 + (3 *^ "xy")_2

vectorFromNonDecList :: (Num k, Eq k, Eq a, Graded a) => [ScalarProduct k a] -> Vector k a

Construct a vector from a list of terms. The list may be infinite.

!!! The grading of terms in the list must be non-descreasing.

Examples:

>>> vectorFromNonDecList [1 *^ 'x', 1 *^ 'y', 1 *^ 'x', 1 *^ 'y']
(2 *^ 'x' + 2 *^ 'y')_1
>>> takeV 10 $ vectorFromNonDecList [1 *^ intToDigits i | i <- [1..]]
(1 *^ [1] + 1 *^ [2] + 1 *^ [3] + 1 *^ [4] + 1 *^ [5] + 1 *^ [6] + 1 *^ [7] + 1 *^ [8] + 1 *^ [9])_1 + (1 *^ [1,0])_2

terms :: Vector k a -> [ScalarProduct k a]

Returns a list of ScalarProduct of a vector.

Examples:

>>> v = Zero &: (1 *^ 'x' +: 1 *^ 'y' +: Zero) &: Empty
>>> terms v
[1 *^ 'x',1 *^ 'y']

termsGraded :: Vector k a -> [[ScalarProduct k a]]

Returns a graded list of ScalarProduct of a vector.

Examples:

>>> v = Zero &: (1 *^ "x" +: 1 *^ "y" +: Zero) &: (1 *^ "xy" +: 1 *^ "yy" +: Zero) &: Empty
>>> termsGraded v
[[],[1 *^ "x",1 *^ "y"],[1 *^ "xy",1 *^ "yy"]]

basisElements :: Vector k a -> [a]

A list of basis elements of a vector.

Examples:

>>> v = vector $ (1 *^ 'x') +: (1 *^ 'y') +: (1 *^ 'x') +: (1 *^ 'y') +: (2 *^ 'z') +: Zero
>>> basisElements v
"xyz"

linear :: (IsVector v, VectorScalar v ~ k, VectorBasis v ~ b) => (a -> v) -> Vector k a -> Vector k b

Takes a function from the basis to a vector space and extends it to a linear map.

!!! the function linear f can receive only finite vectors.

Examples:

>>> f1 x = 1 *^ (x + 1)
>>> v1 = vector $ (1 *^ 1) +: (1 *^ 2) +: (1 *^ 3) +: (1 *^ 4) +: Zero
>>> linear f1 v1
(1 *^ 2 + 1 *^ 3 + 1 *^ 4 + 1 *^ 5)_1
>>> f2 x = vector $ (1 *^ x) +: (1 *^ (x + 1)) +: Zero
>>> v2 = vector $ (1 *^ 1) +: (1 *^ 2) +: (1 *^ 3) +: (1 *^ 4) +: Zero
>>> linear f2 v1
(1 *^ 1 + 2 *^ 2 + 2 *^ 3 + 2 *^ 4 + 1 *^ 5)_1

linearGraded :: (IsVector v, VectorScalar v ~ k, VectorBasis v ~ b) => (a -> v) -> Vector k a -> Vector k b

Takes a function from the basis to a vector space and extends it to a linear map.

!!! f must respect the grading.

Examples:

>>> f (x:xs) = 1 *^ (2:xs)
>>> v1 = vectorFromList [1*^[1], 1*^ [1,1]]
>>> linearGraded f v1
(1 *^ [2])_1 + (1 *^ [2,1])_2
>>> v2 = vectorFromNonDecList [i*^(replicate i 1) | i <- [1..]]
>>> takeV 10 $ linearGraded f v2
(1 *^ [2])_1 + (2 *^ [2,1])_2 + (3 *^ [2,1,1])_3 + (4 *^ [2,1,1,1])_4 + (5 *^ [2,1,1,1,1])_5 + (6 *^ [2,1,1,1,1,1])_6 + (7 *^ [2,1,1,1,1,1,1])_7 + (8 *^ [2,1,1,1,1,1,1,1])_8 + (9 *^ [2,1,1,1,1,1,1,1,1])_9 + (10 *^ [2,1,1,1,1,1,1,1,1,1])_10
>>> v3 = vectorFromNonDecList [1 *^ intToDigits i | i <- [0..]]
>>> g x = 1 *^ (1:x)
>>> linearGraded g v3
*** Exception: Grading mismatch in a list of terms
...

linearNonDec :: (IsVector v, VectorScalar v ~ k, VectorBasis v ~ b) => (a -> v) -> Vector k a -> Vector k b

Takes a function from the basis to a vector space and extends it to a linear map.

!!! the function f must be non-decreasing with respect to the grading, that is, (grading b1) <= (grading b2) implies (min $ grading $ f b1) <= (min $ grading $ f b2), where min is the minimum of the grading of the terms in the image of f.

Examples:

>>> f x = 1 *^ (2:x)
>>> v1 = vectorFromList [1*^[1], 1*^ [1,1]]
>>> linearNonDec f v1
(1 *^ [2,1])_2 + (1 *^ [2,1,1])_3
>>> v2 = vectorFromNonDecList [i*^(replicate i 1) | i <- [1..]]
>>> takeV 10 $ linearNonDec f v2
(1 *^ [2,1])_2 + (2 *^ [2,1,1])_3 + (3 *^ [2,1,1,1])_4 + (4 *^ [2,1,1,1,1])_5 + (5 *^ [2,1,1,1,1,1])_6 + (6 *^ [2,1,1,1,1,1,1])_7 + (7 *^ [2,1,1,1,1,1,1,1])_8 + (8 *^ [2,1,1,1,1,1,1,1,1])_9 + (9 *^ [2,1,1,1,1,1,1,1,1,1])_10 + (10 *^ [2,1,1,1,1,1,1,1,1,1,1])_11
>>> v3 = vectorFromNonDecList [1 *^ intToDigits i | i <- [1..]]
>>> g (x:xs) = if x < 9 then 1 *^ (x + 1:xs) else 1 *^ (1:x:xs)
>>> takeV 10 $ linearNonDec g v3
(1 *^ [2] + 1 *^ [3] + 1 *^ [4] + 1 *^ [5] + 1 *^ [6] + 1 *^ [7] + 1 *^ [8] + 1 *^ [9])_1 + (1 *^ [1,9] + 1 *^ [2,0])_2

bilinear :: (IsVector v, Eq a, Eq b, Graded a, Graded b, VectorScalar v ~ k, VectorBasis v ~ c) => (a -> b -> v) -> Vector k a -> Vector k b -> Vector k c

Takes a function with two arguments and extends it to a bilinear map.

!!! The function bilinear f accepts only finite vectors.

Examples:

>>> f a b = 1 *^ [a + b]
>>> v1 = vector $ (1 *^ 1) +: (1 *^ 2) +: (1 *^ 3) +: (1 *^ 4) +: Zero
>>> v2 = vector $ (1 *^ 1) +: (1 *^ 2) +: (1 *^ 3) +: (1 *^ 4) +: Zero
>>> bilinear f v1 v2
(1 *^ [2] + 2 *^ [3] + 3 *^ [4] + 4 *^ [5] + 3 *^ [6] + 2 *^ [7] + 1 *^ [8])_1

bilinearGraded :: (IsVector v, Eq a, Eq b, Graded a, Graded b, VectorScalar v ~ k, VectorBasis v ~ c) => (a -> b -> v) -> Vector k a -> Vector k b -> Vector k c

Takes a function with two arguments and extends it to a bilinear map.

!!! The function f must preserve the grading, that is, grading (f a b) = (grading a) + (grading b).

Examples:

>>> f1 a b = 1 *^ [a, b]
>>> f2 a b = 1 *^ [a + b]
>>> v1 = vector $ (1 *^ 1) +: (1 *^ 2) +: (1 *^ 3) +: (1 *^ 4) +: Zero
>>> v2 = vector $ (1 *^ 1) +: (1 *^ 2) +: (1 *^ 3) +: (1 *^ 4) +: Zero
>>> bilinearGraded f1 v1 v2
(1 *^ [1,1] + 1 *^ [2,1] + 1 *^ [1,2] + 1 *^ [3,1] + 1 *^ [2,2] + 1 *^ [1,3] + 1 *^ [4,1] + 1 *^ [3,2] + 1 *^ [2,3] + 1 *^ [1,4] + 1 *^ [4,2] + 1 *^ [3,3] + 1 *^ [2,4] + 1 *^ [4,3] + 1 *^ [3,4] + 1 *^ [4,4])_2
>>> bilinearGraded f2 v1 v2
*** Exception: Grading mismatch in a list of terms
...

bilinearNonDec :: (IsVector v, Eq a, Eq b, Graded a, Graded b, VectorScalar v ~ k, VectorBasis v ~ c) => (a -> b -> v) -> Vector k a -> Vector k b -> Vector k c

Takes a function with two arguments and extends it to a bilinear map.

!!! The function f must be non-decreasing with respect to the grading, that is, if (grading a) + (grading b) > (grading c) + (grading d), then grading (f a b) >= grading (f c d).

Examples:

>>> f a b = 1 *^ [a + b]
>>> v1 = vector $ (1 *^ 1) +: (1 *^ 2) +: (1 *^ 3) +: (1 *^ 4) +: Zero
>>> v2 = vector $ (1 *^ 1) +: (1 *^ 2) +: (1 *^ 3) +: (1 *^ 4) +: Zero
>>> bilinearNonDec f v1 v2
(1 *^ [2] + 2 *^ [3] + 3 *^ [4] + 4 *^ [5] + 3 *^ [6] + 2 *^ [7] + 1 *^ [8])_1

linear2 :: (IsVector v1, IsVector v2, IsVector c, IsVector d, Eq a, Eq b, Eq c, Eq d, Graded a, Graded b, Graded c, Graded d, VectorScalar c ~ k, VectorScalar d ~ k, VectorScalar v1 ~ k, VectorScalar v2 ~ k, VectorBasis v1 ~ c, VectorBasis v2 ~ d) => (a -> v1, b -> v2) -> Vector k (a, b) -> Vector k (c, d)

Takes a tuple of two functions and extends it to a linear map.

Examples:

>>> f1 a = 1 *^ (a ++ "1")
>>> f2 a = 1 *^ (a ++ "2")
>>> v = vector $ 1 *^ ("x", "x") +: 1 *^ ("y", "y") +: 1 *^ ("z", "z") +: 1 *^ ("w", "w") +: Zero
>>> linear2 (f1, f2) v
(1 *^ ("x1","x2") + 1 *^ ("y1","y2") + 1 *^ ("z1","z2") + 1 *^ ("w1","w2"))_4

linear2Graded :: (IsVector v1, IsVector v2, IsVector c, IsVector d, Eq a, Eq b, Eq c, Eq d, Graded a, Graded b, Graded c, Graded d, VectorScalar v1 ~ k, VectorScalar v2 ~ k, VectorScalar c ~ k, VectorScalar d ~ k, VectorBasis v1 ~ c, VectorBasis v2 ~ d) => (a -> v1, b -> v2) -> Vector k (a, b) -> Vector k (c, d)

Takes a tuple of two functions and extends it to a linear map.

!!! The functions must preserve the grading.

Examples:

>>> f1 a = 1 *^ (a + 1) :: ScalarProduct Integer Integer
>>> f2 a = 1 *^ (a + 2) :: ScalarProduct Integer Integer
>>> v = vector $ 1 *^ (1, 1) +: 1 *^ (2, 2) +: 1 *^ (3, 3) +: 1 *^ (4, 4) +: Zero
>>> linear2Graded (f1, f2) v
(1 *^ (2,3) + 1 *^ (3,4) + 1 *^ (4,5) + 1 *^ (5,6))_2

linear2NonDec :: (IsVector v1, IsVector v2, IsVector c, IsVector d, Eq a, Eq b, Eq c, Eq d, Graded a, Graded b, Graded c, Graded d, VectorScalar v1 ~ k, VectorScalar v2 ~ k, VectorScalar c ~ k, VectorScalar d ~ k, VectorBasis v1 ~ c, VectorBasis v2 ~ d) => (a -> v1, b -> v2) -> Vector k (a, b) -> Vector k (c, d)

Takes a tuple of two functions and extends it to a linear map.

!!! The functions must be nnon-decreasing with respect to the grading.

Examples:

>>> f1 a = 1 *^ (a + 1) :: ScalarProduct Integer Integer
>>> f2 a = 1 *^ (a + 2) :: ScalarProduct Integer Integer
>>> v = vector $ 1 *^ (1, 1) +: 1 *^ (2, 2) +: 1 *^ (3, 3) +: 1 *^ (4, 4) +: Zero
>>> linear2NonDec (f1, f2) v
(1 *^ (2,3) + 1 *^ (3,4) + 1 *^ (4,5) + 1 *^ (5,6))_2

renormalize :: (Num k2, Eq k2, Eq a, Graded a) => (k1 -> a -> k2) -> Vector k1 a -> Vector k2 a

Change the coefficients in a vector using a function f that takes the scalar and the basis element of a term and returns a new scalar.

Examples:

>>> f s x = s + 1
>>> v = vector $ (1 *^ 'x') +: (2 *^ 'y') +: (3 *^ 'z') +: (4 *^ 'w') +: Zero
>>> renormalize f v
(2 *^ 'x' + 3 *^ 'y' + 4 *^ 'z' + 5 *^ 'w')_1

scale :: (Num k, Eq k, Eq a, Graded a) => k -> Vector k a -> Vector k a

Scale a vector by a scalar.

Examples:

>>> v = vector $ (1 *^ 'x') +: (2 *^ 'y') +: (3 *^ 'z') +: (4 *^ 'w') +: Zero
>>> scale 2 v
(2 *^ 'x' + 4 *^ 'y' + 6 *^ 'z' + 8 *^ 'w')_1

rescale :: (Num k, Eq k, Eq a, Graded a) => (a -> k) -> Vector k a -> Vector k a

Scale a vector by a scalar.

Examples:

>>> v = vectorFromList [1 *^ "x", 2 *^ "xy", 3 *^ "xyz"]
>>> rescale length v
(1 *^ "x")_1 + (4 *^ "xy")_2 + (9 *^ "xyz")_3

changeScalars :: (Num k1, Eq k1, Num k2, Eq k2, Eq a, Graded a) => (k1 -> k2) -> Vector k1 a -> Vector k2 a

Change the scalar of a vector.

Examples:

>>> f s = s + 1
>>> v = vector $ (1 *^ 'x') +: (2 *^ 'y') +: (3 *^ 'z') +: (4 *^ 'w') +: Zero
>>> changeScalars f v
(2 *^ 'x' + 3 *^ 'y' + 4 *^ 'z' + 5 *^ 'w')_1

functional :: (Num k, Eq k, Eq a) => (a -> k) -> Vector k a -> k

Extends a function f that maps basis elements to scalars to a linear functional.

!!! The function functional f accepts only finite vectors.

Examples:

>>> f x = fromInteger $ grading x
>>> v = vectorFromList [1 *^ "x", 2 *^ "y", 3 *^ "z", 4 *^ "ww"]
>>> functional f v
14

lengthV :: Vector k a -> Int

The length of a vector is the number of terms in it.

Examples:

>>> v = vector $ (1 *^ 'x') +: (1 *^ 'y') +: (1 *^ 'z') +: (1 *^ 'w') +: Zero
>>> lengthV v
4

Properties:

lengthV v == length (terms v :: [ScalarProduct Integer Char])

takeWhileV :: (Num k, Eq k, Eq a, Graded a) => (ScalarProduct k a -> Bool) -> Vector k a -> Vector k a

Take terms from a vector until the first term that does not satisfy the condition given by f.

Examples:

>>> f1 (i :*^ j) = j < 3
>>> v1 = vector $ (1 *^ 1) +: (1 *^ 2) +: (1 *^ 3) +: (1 *^ 4) +: Zero
>>> takeWhileV f1 v1
(1 *^ 1 + 1 *^ 2)_1
>>> f2 (i :*^ [j]) = j < 5
>>> v2 = vectorFromNonDecList [1 *^ intToDigits i | i <- [1..]]
>>> takeWhileV f2 v2
(1 *^ [1] + 1 *^ [2] + 1 *^ [3] + 1 *^ [4])_1

Properties:

takeWhileV (\_ -> True) v == (v :: Vector Integer Char)

takeWhileV (\_ -> False) v == (mempty :: Vector Integer Char)

filterV :: (Num k, Eq k, Eq a, Graded a) => (ScalarProduct k a -> Bool) -> Vector k a -> Vector k a

Filter terms from a vector that satisfy the condition given by f.

Examples:

>>> f (_ :*^ (j:_)) = j `mod` 3 == 0
>>> v = vectorFromNonDecList [1 *^ intToDigits i | i <- [1..]]
>>> takeV 10 $ filterV f v
(1 *^ [3] + 1 *^ [6] + 1 *^ [9])_1 + (1 *^ [3,0] + 1 *^ [3,1] + 1 *^ [3,2] + 1 *^ [3,3] + 1 *^ [3,4] + 1 *^ [3,5] + 1 *^ [3,6])_2

Properties:

filterV (\_ -> True) v == (v :: Vector Integer Char)

filterV (\_ -> False) v == (mempty :: Vector Integer Char)

takeV :: (Num k, Eq k, Eq a, Graded a) => Int -> Vector k a -> Vector k a

Take the first n terms from a vector.

Examples:

>>> v = vectorFromNonDecList [1 *^ intToDigits i | i <- [1..]]
>>> takeV 10 v
(1 *^ [1] + 1 *^ [2] + 1 *^ [3] + 1 *^ [4] + 1 *^ [5] + 1 *^ [6] + 1 *^ [7] + 1 *^ [8] + 1 *^ [9])_1 + (1 *^ [1,0])_2

Properties:

takeV (lengthV v) v == (v :: Vector Integer Char)

takeV 0 v == (mempty :: Vector Integer Char)

data Sum k a

A sum is assumed to have a finite number of terms and a grading associated to it.

Constructors

Zero

fromListS :: (Num k, Eq k, Graded a, Eq a) => [ScalarProduct k a] -> Sum k a

Construct a sum from a list of terms with the same grading. The list must be finite.

Examples:

>>> l = [1 *^ 'x', 1 *^ 'y', 1 *^ 'z', 1 *^ 'x', 1 *^ 'y', (-2) *^ 'z'] :: [ScalarProduct Integer Char]
>>> fromListS l
2 *^ 'x' + 2 *^ 'y' + -1 *^ 'z'

Properties:

\l -> lengthS (fromListS l) <= length (l :: [ScalarProduct Integer Char])

toListS :: Sum k a -> [ScalarProduct k a]

Construct a list of terms from a sum.

Examples:

>>> s = 2 *^ 'x' +: 2 *^ 'y' +: Zero :: Sum Integer Char
>>> toListS s
[2 *^ 'x',2 *^ 'y']

Properties:

\s -> length (toListS s) == lengthS (s :: Sum Integer Char)

pattern (:+) :: ScalarProduct k a -> Sum k a -> Sum k a

(+:) :: (Num k, Eq k, Eq a, Graded a) => ScalarProduct k a -> Sum k a -> Sum k a infixr 6

Adds a term to a sum ensuring that every term in the sum has a unique basis element and the same grading.

Examples:

>>> (1 *^ 'x') +: (1 *^ 'x') +: (1 *^ 'y') +: Zero
2 *^ 'x' + 1 *^ 'y'
>>> (1 *^ 'z') +: (1 *^ 'x') +: (1 *^ 'y') +: Zero
1 *^ 'z' + 1 *^ 'x' + 1 *^ 'y'

Properties:

\t l -> (lengthS $ t +: l) - 1 <= lengthS (l :: Sum Integer Char)

data Vector k a

A power series can have infinite number of terms.

Constructors

Empty

fromListV :: (Num k, Eq k, Eq a, Graded a) => [Sum k a] -> Vector k a

Construct a vector from a list of sums where position i of the list is holds a sum with grading i or 0. The list may be infinite.

Examples:

>>> s1 = (1 *^ "") +: Zero
>>> s2 = (1 *^ "x") +: (1 *^ "y") +: (2 *^ "z") +: Zero
>>> s3 = (1 *^ "xx") +: (1 *^ "xy") +: (2 *^ "zz") +: Zero
>>> fromListV [s1, s2, s3]
(1 *^ "")_0 + (1 *^ "x" + 1 *^ "y" + 2 *^ "z")_1 + (1 *^ "xx" + 1 *^ "xy" + 2 *^ "zz")_2

>>> fromListV [s2, s3]
*** Exception: Grading mismatch in a list of terms
...
>>> fromListV [s1, s3]
(1 *^ "")_0*** Exception: Grading mismatch in a list of terms
...
>>> fromListV [s1, Zero, s3, Zero]
(1 *^ "")_0 + (1 *^ "xx" + 1 *^ "xy" + 2 *^ "zz")_2

toListV :: Vector k a -> [Sum k a]

Return a list of sums from a vector.

Examples:

>>> v = Zero &: (1 *^ "x") +: (1 *^ "y") +: Zero &: (1 *^ "xy") +: (1 *^ "yx") +: (2 *^ "zz") +: Zero &: Empty
>>> toListV v
[0,1 *^ "x" + 1 *^ "y",1 *^ "xy" + 1 *^ "yx" + 2 *^ "zz"]